feat: add chat interface components and settings management
- Implemented EmptyState component for chat interface with suggestions.
- Created LoadingState component to indicate loading status.
- Added ShareTokenModal for sharing chat links.
- Developed CodeBlock component for syntax highlighting in markdown.
- Introduced CustomMarkdown parser for rendering markdown content.
- Added ApiKeyTab for managing API keys with validation.
- Created AppearanceTab for future theme customization.
- Implemented DataTab for managing chat data with delete functionality.
- Developed GeneralTab for configuring AI model and search preferences.
- Added Notification component for displaying error and success messages.
- Created SettingsModal for managing application settings.
- Implemented hooks for API key management and notifications.
- Defined TypeScript interfaces for markdown and settings.
