+++ src/types/markdown.ts
@@ -1,0 +1,10 @@
+export interface MarkdownParserProps {
+  content: string;
+  className?: string;
+}
+
+export interface CodeBlockProps {
+  code: string;
+  language: string;
+  index: number;
+}
